Add the `aeron` build tree to both crates
Because `cargo publish` won't look at parent directories, we can't use
`../aeron` to share the build tree between crates. Instead, we need to
keep multiple copies so that each crate can be used independently.

We'd typically prefer a build structure where `libaeron-sys` just
includes the media driver as a feature, but because the client and
driver are separate targets in Cmake (and thus separate libraries),
we leave them as separate crates here.